redis/requirements.txt
pydantic
redis-om
kafka-python